pyproject.toml
[project]
name = "cv-match-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.1",
    "langchain-anthropic>=1.4.1",
    "langgraph>=1.1.9",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "streamlit>=1.56.0",
    "uvicorn>=0.46.0",
]
